src-python/requirements.txt
onnxruntime-gpu
fastapi
uvicorn
librosa
soundfile
pydub
numpy
requests
python-multipart
sse-starlette
